Handle user's import error when module in test spec.
ImportError was already raised for test spec like "pkg1.pkg2" when
a module of pkg2 had a failing import statement. But it was not raised
when the module was part of the test spec like in
"pkg1.pkg2.module_with_import_error".

This patch fix this. It helps the user to get the line where the
ImportError happen.