StopWatch returns zero for most constants when not started.
If the stopwatch is not started that probably mean there were no test
at all. Thus, total_split_time, mean_split_time and total_time should
be equal to zero.

It eases the caller code that does not have to handle an exception in
such a case. If it needs to do a special treatment, it can still call
'is_started'.